Changed the way barycentric interpolation is done.
Before, in the original Triangle, barycentric interpolation for the
attributes of a point inserted at the circumcenter of a triangle was
done using normal barycentric coordinates. However if the triangle for
which we found the circumcenter was very obtuse then the point in the
circumcenter was very far away from the triangle and therefore
interpolation of the atributes was subject to the inherent noise in the
atributes (in our case distance to the sensor). The new point would
therefore end up with a negative distance or a very large one.
The solution is to take barycentric coordinates not from the triangle in
which we found the circumcenter but rather the triangle on which the
newvertex lies on. The function insertvertex finds that triangle and
splits it into 3 smaller ones. That's the triangle we use for
interpolation since we know the newvertex will be somewhere inside of
it.